requirements.txt
aiohttp==0.21.6
cffi==1.7.0
chardet==2.3.0
discord.py==0.11.0
pycparser==2.14
PyNaCl==1.0.1
six==1.10.0
websockets==3.1
